BackendDjango/backend/requirements.txt
asgiref==3.5.0
async-generator==1.10
attrs==21.4.0
beautifulsoup4==4.10.0
bson==0.5.8
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
coreapi==2.3.3
coreschema==0.0.4
cryptography==36.0.2
dataclasses==0.6
Django==4.0.3
django-cors-headers==3.11.0
djangorestframework==3.13.1
djongo==1.3.6
drf-yasg==1.20.0
h11==0.13.0
idna==3.3
inflection==0.5.1
itypes==1.2.0
Jinja2==3.0.3
joblib==1.1.0
MarkupSafe==2.1.1
mysqlclient==2.1.0
numpy==1.22.2
outcome==1.1.0
packaging==21.3
pandas==1.4.1
pycparser==2.21
pymongo==3.12.3
pyOpenSSL==22.0.0
pyparsing==3.0.7
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2022.1
requests==2.27.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
scikit-learn==1.0.2
scipy==1.8.0
selenium==4.1.3
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
soupsieve==2.3.1
sqlparse==0.2.4
threadpoolctl==3.1.0
trio==0.20.0
trio-websocket==0.9.2
tzdata==2022.1
uritemplate==4.1.1
urllib3==1.26.9
wsproto==1.1.0
